+++ code/paul-eduard/include/merge-sort.h
@@ -3,8 +3,9 @@
 #ifndef CODE_PAUL_EDUARD_INCLUDE_MERGE_SORT_H_
 #define CODE_PAUL_EDUARD_INCLUDE_MERGE_SORT_H_
 
+#include <stdint.h>
 #include <vector>
 
-int Merge_Sort(std::vector<int> *mas, unsigned int l, unsigned int r);
+int Merge_Sort(std::vector<int64_t> *mas, unsigned int l, unsigned int r);
 
 #endif  // CODE_PAUL_EDUARD_INCLUDE_MERGE_SORT_H_
