+++ code/paul-eduard/src/merge-sort.cxx
@@ -1,16 +1,17 @@
 // Copyright 2014 Paul Eduard
 
 #include "include/merge-sort.h"
+#include <stdint.h>
 #include <vector>
 #include <climits>
 
-static void Merge(std::vector<int> *mas, unsigned int left,
+static void Merge(std::vector<int64_t> *mas, unsigned int left,
     unsigned int m, unsigned int right) {
 
     unsigned int n1 = m-left+1;
     unsigned int n2 = right-m;
-    std::vector<int> L(n1);
-    std::vector<int> R(n2);
+    std::vector<int64_t> L(n1);
+    std::vector<int64_t> R(n2);
     for (unsigned int i = 0; i < n1; ++i) {
         L[i] = (*mas)[left+i];
     }
@@ -31,7 +32,7 @@ static void Merge(std::vector<int> *mas, unsigned int left,
     while (j < n2) (*mas)[k++] = R[j++];
 }
 
-int Merge_Sort(std::vector<int> *mas, unsigned int l, unsigned int r) {
+int Merge_Sort(std::vector<int64_t> *mas, unsigned int l, unsigned int r) {
     if ((*mas).size() == 0) return -1;
     if (l >= r) return 0;
     unsigned int middle = (l+r)/2;
